+++ src/components/Fieldset/Fieldset.tsx
@@ -1,6 +1,7 @@
 import classNames from "classnames";
 import PropTypes from "prop-types";
 import React from "react";
+import { nullValuesAsUndefined } from "null-as-undefined";
 
 import { Attributes, DataAttributes } from "../../helpers/Attributes";
 
@@ -13,26 +14,37 @@ import "lbh-frontend/lbh/components/lbh-fieldset/_fieldset.scss";
  *
  * @noInheritDoc
  */
-export interface FieldsetProps extends React.AriaAttributes, DataAttributes {
-  id?: string;
-  className?: string;
-  role?: string;
+export interface FieldsetProps extends FieldsetPropsWithoutChildren {
+  children: React.ReactNode;
+}
+
+export interface FieldsetPropsWithoutChildren
+  extends React.AriaAttributes,
+    DataAttributes {
+  id?: string | null;
+  className?: string | null;
+  role?: string | null;
   /**
    * The legend for the fieldset. You should pass in a {@link FieldsetLegend}
    * component.
    */
-  legend?: React.ReactNode;
-  describedBy?: string;
-  children: React.ReactNode;
+  legend?: React.ReactNode | null;
+  describedBy?: string | null;
 }
-
 /**
  * A component used to group elements within a form.
  */
 export const Fieldset: React.FunctionComponent<FieldsetProps> = (
   props: FieldsetProps
 ): React.ReactElement => {
-  const { id, className, legend, role, describedBy, children } = props;
+  const {
+    id,
+    className,
+    legend,
+    role,
+    describedBy,
+    children
+  } = nullValuesAsUndefined(props);
 
   const extraAttributes = Attributes.ariaAndData(props);
 
@@ -50,11 +62,15 @@ export const Fieldset: React.FunctionComponent<FieldsetProps> = (
   );
 };
 
-Fieldset.propTypes = {
+export const fieldsetWithoutChildrenPropTypes: PropTypes.ValidationMap<FieldsetPropsWithoutChildren> = {
   id: PropTypes.string,
   className: PropTypes.string,
   role: PropTypes.string,
   legend: PropTypes.element,
-  describedBy: PropTypes.string,
+  describedBy: PropTypes.string
+};
+
+Fieldset.propTypes = {
+  ...fieldsetWithoutChildrenPropTypes,
   children: PropTypes.node.isRequired
 };